apps/termene/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.8
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
